FLUX: extract FluxPipeline from __init__.py to flux.py
